Enh: replace deprecated mt_spec calls in preproc.quality, move mt_spec to old
